moved common logic to abstract class and made power generators derive from it
